Moves pagination tests
