+++ packages/pagination/__tests__/android/pagination-icons.native.test.js
