+++ packages/pagination/__tests__/android/pagination-wrapper.native.test.js
@@ -1,4 +1,4 @@
 import tests from "../test-wrapper-helper";
-import pagination from "../../pagination-wrapper";
+import pagination from "../../src/pagination-wrapper";
 
 describe("Pagination Wrapper on Android", tests(pagination));
