+++ packages/pagination/__tests__/ios/pagination.native.test.js
@@ -1,4 +1,4 @@
 import tests from "../test-helper";
-import Pagination from "../../pagination";
+import Pagination from "../../src/pagination";
 
 describe("Pagination on iOS", tests(Pagination));
