+++ packages/pagination/__tests__/web/pagination.web.test.js
@@ -1,5 +1,5 @@
 import "jest-styled-components";
 import tests from "../test-helper";
-import Pagination from "../../pagination";
+import Pagination from "../../src/pagination";
 
 describe("Pagination on Web", tests(Pagination));
